fixed pagination bug, fixed units not being able to be floats, added shadow modal exit,

+++ modules/audit.php
@@ -51,7 +51,7 @@
 					'listing' => $return,
 					'count' => intval( $count ),
 					'limit' => $limit,
-					'currentPage' => ++$page
+					'currentPage' => (int)++$page
 				);
 			}
 		}
